Fix import of stylesheets and loading of images
Fixed the import of the stylesheets. Now the screen is styled according to the stylesheet. The import of the font does not work at the moment. Also, the screen is too big to be displayed correctly if content of screens is scaled by 125% (default setting on Windows, can be set to 100%). Images are now loading correctly.