Game now also takes a QuestionController as parameter
Game class now also takes a QuestionController as a parameter. Not sure if this is the best way of hadling it, but otherwise the server would proably have to send API requests to itself, which would not be optimal. Now the Game can simply call the methods in the QuestionController.